Removed annotation.py and renamed annotation_v2.py
